[REFACTOR#1560] noRippleClickable 공통 유틸리티로 이동 및 통합
