+++ core/common/src/main/java/org/sopt/official/common/util/ClickableExt.kt
@@ -1,7 +1,11 @@
 package org.sopt.official.common.util
 
 import android.os.SystemClock
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.composed
 import androidx.compose.ui.input.pointer.PointerEvent
 import androidx.compose.ui.input.pointer.PointerEventPass
 import androidx.compose.ui.input.pointer.PointerEventType
@@ -58,3 +62,15 @@ private class ThrottledNoRippleClickableNode(
 
     override fun onCancelPointerInput() = Unit
 }
+
+inline fun Modifier.noRippleClickable(crossinline onClick: () -> Unit): Modifier =
+    composed {
+        then(
+            Modifier.clickable(
+                indication = null,
+                interactionSource = remember { MutableInteractionSource() },
+            ) {
+                onClick()
+            },
+        )
+    }
