+++ feature/soptamp/src/main/java/org/sopt/official/stamp/designsystem/component/button/SoptampIconButton.kt
@@ -36,7 +36,7 @@ import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.tooling.preview.Preview
 import org.sopt.official.designsystem.SoptTheme
 import org.sopt.official.stamp.R
-import org.sopt.official.stamp.designsystem.component.util.noRippleClickable
+import org.sopt.official.common.util.noRippleClickable
 
 /**
  * 앱 디자인 시스템 Icon Button 입니다.
