+++ feature/soptamp/src/main/java/org/sopt/official/stamp/designsystem/component/dialog/DoubleOptionDialog.kt
@@ -43,7 +43,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import org.sopt.official.designsystem.SoptTheme
-import org.sopt.official.stamp.designsystem.component.util.noRippleClickable
+import org.sopt.official.common.util.noRippleClickable
 import org.sopt.official.stamp.util.DefaultPreview
 
 @Composable
