+++ feature/soptamp/src/main/java/org/sopt/official/stamp/designsystem/component/mission/MissionComponent.kt
@@ -43,7 +43,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import org.sopt.official.designsystem.SoptTheme
 import org.sopt.official.domain.soptamp.MissionLevel
-import org.sopt.official.stamp.designsystem.component.util.noRippleClickable
+import org.sopt.official.common.util.noRippleClickable
 import org.sopt.official.stamp.feature.mission.model.MissionUiModel
 
 @Composable
