+++ feature/soptamp/src/main/java/org/sopt/official/stamp/feature/mission/detail/component/ImageModal.kt
@@ -52,7 +52,7 @@ import coil.request.ImageRequest
 import coil.transform.RoundedCornersTransformation
 import org.sopt.official.designsystem.SoptTheme
 import org.sopt.official.stamp.R
-import org.sopt.official.stamp.designsystem.component.util.noRippleClickable
+import org.sopt.official.common.util.noRippleClickable
 
 @Composable
 fun ImageModal(
