+++ feature/soptamp/src/main/java/org/sopt/official/stamp/feature/ranking/TopRankerItem.kt
@@ -50,7 +50,7 @@ import androidx.compose.ui.unit.dp
 import org.sopt.official.designsystem.SoptTheme
 import org.sopt.official.stamp.R
 import org.sopt.official.stamp.designsystem.component.button.SoptampIconButton
-import org.sopt.official.stamp.designsystem.component.util.noRippleClickable
+import org.sopt.official.common.util.noRippleClickable
 import org.sopt.official.stamp.designsystem.style.Pink300
 import org.sopt.official.stamp.feature.ranking.model.RankerUiModel
